wpp-connect-api/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic-settings
python-dotenv
alembic
httpx
aioboto3
aiofiles
